Implemented Swiper carousel and fixed layout issues in ProductGrid
