+++ src/app/shop/components/alsolike.tsx
@@ -3,6 +3,7 @@ import Wrapper from "@/layout/wrapper";
 import { products } from "@/ui/products/_data";
 import ProductGrid from "@/ui/products/ProductGrid";
 import React from "react";
+import ProductRecommedation from "./ui/ProductRecommedation";
 
 const AlsoLike = () => {
     const limitedProducts = products.slice(0, 5);
@@ -18,7 +19,8 @@ const AlsoLike = () => {
         <h1 className="uppercase text-primary-2/85 font-bold text-[2rem] leading-9 pt-[3rem] flex-1 mb-[2.5rem]">
           You may also like
         </h1>
-        <ProductGrid products={limitedProducts} />
+        {/* <ProductGrid products={limitedProducts} /> */}
+        <ProductRecommedation products={limitedProducts}/>
       </div>
      
     </Wrapper>
